+++ printcetnrum_web.client/printcetnrum_web.client.esproj
@@ -13,6 +13,5 @@
   <ItemGroup>
     <Folder Include="src\app\guards\" />
     <Folder Include="src\app\interceptors\" />
-    <Folder Include="src\app\orders\" />
   </ItemGroup>
 </Project>
