--- printcetnrum_web.client/src/app/about-page/about-page.component.ts
@@ -1,15 +1,0 @@
-import { Component } from '@angular/core';
-
-
-@Component({
-  selector: 'app-about-page',
-  templateUrl: './about-page.component.html',
-  styleUrl: './about-page.component.css'
-})
-export class AboutPageComponent {
-  
-}
-
-//function HostListener(arg0: string): (target: AboutPageComponent, propertyKey: "onWindowScroll", descriptor: TypedPropertyDescriptor<() => void>) => void | TypedPropertyDescriptor<() => void> {
-//    throw new Error('Function not implemented.');
-//}
